Report individual findings
This PR changes the way vulnerabilities are reported.
Each finding is now reported individually, rather than reporting
and aggregation of all findings with only the top-level file on
disk.

For example, a vulnerable jar nested inside an archive will now be
reported with the vulnerability findings, rather than reporting on
the archive with an aggregation of all findings from within it.

The path reported with a vulnerability finding is the full path
to a finding with archive layers delimited by a "!".
i.e. /path/to/archive!path/to/finding.jar shows that an archive
 at /path/to/archive contained a vulnerable jar at
 path/to/finding.jar within it.