Skip all class checks when not .class
When checking a file within an archive, if the path has no .class
suffix, all class-based checks will be skipped.

Similarly, if the path does have a .class suffix, there will be no
attempt to check the filename against the log4j jar regex, which
would be impossible to match.